+++ pyproject.toml
@@ -38,6 +38,7 @@ isort = "^5.9.3"
 pytest-cov = "^3.0.0"
 commitizen = "^2.20.0"
 anyio = {extras = ["trio"], version = "^3.3.4"}
+unasync-cli = {url = "https://github.com/leynier/unasync-cli/releases/download/v0.0.5/unasync-cli-0.0.5.tar.gz"}
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
